microservices/reco-builder/requirements.txt
flask==3.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
minio==7.2.0
python-multipart==0.0.6
transformers==4.38.2
pika==1.3.2
py-eureka-client==0.11.1
